refactor(dir_graph): split schema accessors into schema_ops.rs (god-file gate)
The Phase-1 `primary_key_for` addition pushed dir_graph/mod.rs to 3002 LoC,
2 over the 3000 default god-file ceiling (test_god_files.py — caught by the
full CI suite, not the targeted local subset). Per the project rule, split
rather than bump the allowlist: move the four schema-definition accessors
(set/get/clear_schema + primary_key_for) into a sibling `schema_ops.rs`
`impl DirGraph` block. Behaviour-identical; mod.rs → 2974 LoC.
